+++ kroxylicious-runtime/src/test/java/io/kroxylicious/proxy/config/SniHostIdentifiesNodeIdentificationStrategyTest.java
@@ -7,11 +7,14 @@
 package io.kroxylicious.proxy.config;
 
 import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
 import io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider.SniRoutingClusterNetworkAddressConfigProvider;
+import io.kroxylicious.proxy.model.VirtualClusterModel;
 import io.kroxylicious.proxy.service.HostPort;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.when;
 
 @SuppressWarnings("removal")
 class SniHostIdentifiesNodeIdentificationStrategyTest {
@@ -20,7 +23,7 @@ class SniHostIdentifiesNodeIdentificationStrategyTest {
 
     @Test
     void providesDefinition() {
-        var strategy = new SniHostIdentifiesNodeIdentificationStrategy(HostPort.parse("boot:1234"), "mybroker-$(nodeId)");
+        var strategy = new SniHostIdentifiesNodeIdentificationStrategy("boot:1234", "mybroker-$(nodeId)");
         var definition = strategy.get();
         assertThat(definition.type()).isEqualTo(SniRoutingClusterNetworkAddressConfigProvider.class.getSimpleName());
     }
@@ -28,10 +31,33 @@ class SniHostIdentifiesNodeIdentificationStrategyTest {
     @Test
     void canConstructProviderFromDefinition() {
         var bootstrap = HostPort.parse("boot:1234");
-        var strategy = new SniHostIdentifiesNodeIdentificationStrategy(bootstrap, "mybroker-$(nodeId)");
+        var strategy = new SniHostIdentifiesNodeIdentificationStrategy(bootstrap.toString(), "mybroker-$(nodeId)");
         var definition = strategy.get();
 
-        var provider = service.build((SniRoutingClusterNetworkAddressConfigProvider.SniRoutingClusterNetworkAddressConfigProviderConfig) definition.config());
+        VirtualClusterModel mock = Mockito.mock(
+                VirtualClusterModel.class);
+        when(mock.getClusterName()).thenReturn("my-cluster");
+        var provider = service.build((SniRoutingClusterNetworkAddressConfigProvider.SniRoutingClusterNetworkAddressConfigProviderConfig) definition.config(),
+                mock);
         assertThat(provider.getClusterBootstrapAddress()).isEqualTo(bootstrap);
+    }
+
+    @Test
+    void canConstructProviderFromDefinitionWithClusterNameReplacementToken() {
+        String virtualClusterName = "my-cluster";
+        var bootstrap = HostPort.parse("boot:1234");
+        var strategy = new SniHostIdentifiesNodeIdentificationStrategy(bootstrap.toString(), "my-broker-$(virtualClusterName)-$(nodeId)");
+        var definition = strategy.get();
+
+        VirtualClusterModel mock = Mockito.mock(
+                VirtualClusterModel.class);
+        when(mock.getClusterName()).thenReturn(virtualClusterName);
+        var provider = service.build((SniRoutingClusterNetworkAddressConfigProvider.SniRoutingClusterNetworkAddressConfigProviderConfig) definition.config(),
+                mock);
+        assertThat(provider.getClusterBootstrapAddress()).isEqualTo(bootstrap);
+        assertThat(provider.getBrokerAddress(1)).isEqualTo(HostPort.parse("my-broker-my-cluster-1:1234"));
+        assertThat(provider.getBrokerIdFromBrokerAddress(HostPort.parse("my-broker-my-cluster-1:1234"))).isEqualTo(1);
+        assertThat(provider.getBrokerIdFromBrokerAddress(HostPort.parse("my-broker-another-cluster-1:1234"))).isNull();
+        assertThat(provider.getAdvertisedBrokerAddress(1)).isEqualTo(HostPort.parse("my-broker-my-cluster-1:1234"));
     }
 }